requirements.txt
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
billiard==4.2.0
cachetools==5.3.2
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
cssselect==1.2.0
cssutils==2.9.0
distro==1.8.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
emails==0.6
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
kombu==5.3.3
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
openai==1.6.1
packaging==23.2
passlib==1.7.4
Pillow==10.1.0
pluggy==1.3.0
premailer==3.10.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.23.2
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.10
websockets==12.0